Set explicit size for most variables

+++ lib/EEPROMStorage/EEPROMBackedArray.h
@@ -1,31 +1,31 @@
 #pragma once
 
 #include "EEPROMStorage.h"
-template <int N, int M >
+template <int16_t N, int16_t M >
 class EEPROMBackedArrayInterface
 {
     public:        
-    virtual int const * getSubArray( int n ) = 0;
-    int getArrayLength() { return M; }
+    virtual int16_t const * getSubArray( int16_t n ) = 0;
+    int16_t getArrayLength() { return M; }
 };
 
-template <int N, int M >
+template <int16_t N, int16_t M >
 class EEPROMBackedArray: public EEPROMBackedArrayInterface<N,M>
 {
     private:
         EEPROMStorageInterface* eeprom;
-        int values[N][M];   
-        const int startIndex;     
+        int16_t values[N][M];   
+        const int16_t startIndex;     
     
     public:
-        EEPROMBackedArray( EEPROMStorageInterface* _eeprom, int _startIndex ):
+        EEPROMBackedArray( EEPROMStorageInterface* _eeprom, int16_t _startIndex ):
             startIndex(_startIndex)
         {
             eeprom = _eeprom;            
             memset(values,0,N*M*sizeof(int));
         }
 
-        void setValue( int n, int m, int value )
+        void setValue( int16_t n, int16_t m, int16_t value )
         {
             if ( n < N && m < M )
                 values[n][m] = value;
@@ -34,15 +34,15 @@ class EEPROMBackedArray: public EEPROMBackedArrayInterface<N,M>
 
         void loadFromEEPROM()
         {
-            eeprom->readIntArrayFromAddress(startIndex,(int*)values,N*M);
+            eeprom->readIntArrayFromAddress(startIndex,(int16_t*)values,N*M);
         }
 
         void storeToEEPROM()
         {
-            eeprom->writeIntArrayToAddress(startIndex,(int*)values,N*M);
+            eeprom->writeIntArrayToAddress(startIndex,(int16_t*)values,N*M);
         }
         
-        virtual int const * getSubArray( int n )
+        virtual int16_t const * getSubArray( int16_t n )
         {            
             return values[n];            
         }
