+++ test/_testEEPROMStorage/testEEPROMBackedStorage.cpp
@@ -14,7 +14,7 @@ void testConstructEEPROMStorage()
 
 void testLoadFromEEPROM()
 {
-    int expected[2][2] = {{1, 2}, {3, 4}};
+    int16_t expected[2][2] = {{1, 2}, {3, 4}};
     MockEEPROMStorage eeprom;
     eeprom.rawData = (int*)expected;
 
@@ -34,7 +34,7 @@ void testSetValue()
 
 void testStoreToEEPROM()
 {
-    int expected[2][2] = {{0, 0}, {0, 0}};
+    int16_t expected[2][2] = {{0, 0}, {0, 0}};
     MockEEPROMStorage eeprom;
     eeprom.rawData = (int*)expected;
     EEPROMBackedArray<2,2> array(&eeprom, 0);    
@@ -47,14 +47,14 @@ void testStoreToEEPROM()
 
 void testGetSubArray()
 {
-    int expected[2][2] = {{1, 2}, {3, 4}};
+    int16_t expected[2][2] = {{1, 2}, {3, 4}};
     MockEEPROMStorage eeprom;
     eeprom.rawData = (int*)expected;
 
     EEPROMBackedArray<2,2> array(&eeprom, 0);    
     array.loadFromEEPROM();
 
-    int const * subArray = array.getSubArray(1);
+    int16_t const * subArray = array.getSubArray(1);
 
     TEST_ASSERT_EQUAL_INT_ARRAY( expected[1], subArray, 2 );           
 }
@@ -62,7 +62,7 @@ void testGetSubArray()
 void testGetArrayLength()
 {
     EEPROMBackedArray<2,5> array(0, 0);    
-    int len = array.getArrayLength();
+    int16_t len = array.getArrayLength();
     
     TEST_ASSERT_EQUAL_INT( 5, len );
 }
